requirements.txt
# Core MCP Server Dependencies
fastmcp>=2.13.0
superlinked>=37.5.0
pandas>=2.2.0
python-dotenv>=1.1.0

# Embedding Model
sentence-transformers>=5.1.0
torch>=2.9.0

# Streamlit Demo Dependencies
streamlit>=1.50.0
claude-agent-sdk>=0.1.5
nest-asyncio>=1.6.0

# Optional: Additional data format support
openpyxl>=3.1.0  # For Excel files
